Exercise 4.21*

+++ controllers/blogs.js
@@ -38,8 +38,21 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!request.token || !decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if(blogToDelete.user.toString() === decodedToken.id.toString())
+  {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  }
+  else{
+    return response.status(401).json({ error: `the user ${decodedToken.username} is not allowed to do this action on this blog` })
+  }
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
